requirements.txt
pandas==2.2.2
openpyxl==3.1.5
requests==2.32.3
pyyaml==6.0.2
flask==3.0.3